refactor: reorganized components based on Atomic Design principles
- Moved ItemList to molecules for better granularity.
- Centralized shared types and kept local types near their components.
- Improved project structure for maintainability and scalability.
