+++ src/components/atoms/Item/Item.types.ts
@@ -1,0 +1,5 @@
+import { Item } from "../../../types/Item.types";
+
+export interface ItemProps {
+  item: Item;
+}
