+++ src/components/atoms/Textarea/Textarea.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TextareaProps } from "../../types/Textarea.types";
+import { TextareaProps } from "./Textarea.types";
 
 export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
   ({ placeholder, id, ...props }, ref) => {
@@ -10,7 +10,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
           className="block py-2.5 px-0 w-full text-sm text-gray-900 bg-transparent border-0 border-b-2 border-gray-300 appearance-none focus:outline-none focus:ring-0 focus:border-blue-600 peer   dark:border-gray-600 dark:focus:border-blue-500"
           id={id}
           placeholder=" "
-          {...props} 
+          {...props}
         />
         <label
           htmlFor={id}
