+++ src/components/molecules/ItemList/ItemList.types.ts
@@ -1,4 +1,4 @@
-import { Item } from "./Item.types";
+import { Item } from "../../../types/Item.types";
 
 export interface ItemListProps {
   items: Item[];
